refactor: 이벤트 패키지 구조 개편
