+++ cherrypic-api/src/main/java/org/cherrypic/domain/album/dto/event/AlbumImagesDeleteEvent.java
@@ -1,4 +1,4 @@
-package org.cherrypic.domain.album.event;
+package org.cherrypic.domain.album.dto.event;
 
 public record AlbumImagesDeleteEvent(Long albumId) {
     public static AlbumImagesDeleteEvent of(Long albumId) {
