+++ cherrypic-api/src/main/java/org/cherrypic/domain/image/dto/event/ImageDeleteEvent.java
@@ -1,4 +1,4 @@
-package org.cherrypic.domain.image.event;
+package org.cherrypic.domain.image.dto.event;
 
 public record ImageDeleteEvent(String imageUrl) {
     public static ImageDeleteEvent of(String imageUrl) {
