+++ cherrypic-api/src/main/java/org/cherrypic/domain/image/dto/event/ImagesDeleteEvent.java
@@ -1,4 +1,4 @@
-package org.cherrypic.domain.image.event;
+package org.cherrypic.domain.image.dto.event;
 
 import java.util.List;
