+++ cherrypic-api/src/test/java/org/cherrypic/album/service/AlbumServiceTest.java
@@ -15,13 +15,13 @@ import org.cherrypic.RedisCleaner;
 import org.cherrypic.album.entity.Album;
 import org.cherrypic.album.entity.InvitationCode;
 import org.cherrypic.album.enums.AlbumType;
+import org.cherrypic.domain.album.dto.event.AlbumDeleteNotificationSendEvent;
+import org.cherrypic.domain.album.dto.event.AlbumImagesDeleteEvent;
 import org.cherrypic.domain.album.dto.request.AlbumCreateRequest;
 import org.cherrypic.domain.album.dto.request.AlbumUpdateRequest;
 import org.cherrypic.domain.album.dto.response.AlbumInfoResponse;
 import org.cherrypic.domain.album.dto.response.AlbumListResponse;
 import org.cherrypic.domain.album.dto.response.InvitationLinkCreateResponse;
-import org.cherrypic.domain.album.event.AlbumDeleteNotificationSendEvent;
-import org.cherrypic.domain.album.event.AlbumImagesDeleteEvent;
 import org.cherrypic.domain.album.exception.AlbumErrorCode;
 import org.cherrypic.domain.album.repository.AlbumRepository;
 import org.cherrypic.domain.album.repository.InvitationCodeRepository;
@@ -29,8 +29,8 @@ import org.cherrypic.domain.album.service.AlbumService;
 import org.cherrypic.domain.event.repository.EventImageRepository;
 import org.cherrypic.domain.event.repository.EventRepository;
 import org.cherrypic.domain.favorites.repository.FavoritesRepository;
-import org.cherrypic.domain.image.event.ImageDeleteEvent;
-import org.cherrypic.domain.image.event.ImagesDeleteEvent;
+import org.cherrypic.domain.image.dto.event.ImageDeleteEvent;
+import org.cherrypic.domain.image.dto.event.ImagesDeleteEvent;
 import org.cherrypic.domain.image.repository.ImageRepository;
 import org.cherrypic.domain.member.repository.MemberRepository;
 import org.cherrypic.domain.notification.repository.NotificationRepository;
