+++ cherrypic-api/src/test/java/org/cherrypic/image/service/ImageServiceTest.java
@@ -15,16 +15,16 @@ import org.cherrypic.domain.album.repository.AlbumRepository;
 import org.cherrypic.domain.event.exception.EventErrorCode;
 import org.cherrypic.domain.event.repository.EventImageRepository;
 import org.cherrypic.domain.event.repository.EventRepository;
+import org.cherrypic.domain.image.dto.event.ImagesDeleteEvent;
 import org.cherrypic.domain.image.dto.request.*;
 import org.cherrypic.domain.image.dto.response.*;
-import org.cherrypic.domain.image.event.ImagesDeleteEvent;
 import org.cherrypic.domain.image.exception.ImageErrorCode;
 import org.cherrypic.domain.image.repository.ImageRepository;
 import org.cherrypic.domain.image.service.ImageService;
 import org.cherrypic.domain.member.repository.MemberRepository;
 import org.cherrypic.domain.participant.repository.ParticipantRepository;
 import org.cherrypic.domain.subscription.repository.SubscriptionRepository;
-import org.cherrypic.domain.tempalbum.event.TempAlbumImagesDeleteEvent;
+import org.cherrypic.domain.tempalbum.dto.event.TempAlbumImagesDeleteEvent;
 import org.cherrypic.domain.tempalbum.exception.TempAlbumErrorCode;
 import org.cherrypic.domain.tempalbum.repository.TempAlbumImageRepository;
 import org.cherrypic.domain.tempalbum.repository.TempAlbumRepository;
